Update and rename model.py to model_with_pT_normalization.py
This model is not the same one that WSO used to get the results detailed in their end of summer presentation (see [1]). This model, in which the pT variables are normalized by total pT, does NOT perform as well. I'm leaving it here for historical reasons, but we probably want to move away from this.

1. https://docs.google.com/presentation/d/1V6UwJSQ6zp3HWdvUhwIrohQ6-u5zRTiAnlsvEhLabRU/edit#slide=id.g5ed33f3b16_0_285