Add files via upload
Condor scripts to use for preprocessing. Have been tested on Brux and work there, should in principle also work on lxplus condor, as lxplus condor nodes, like Brux condor nodes, are able to interact with the ordinary workspaces, but have not been tested at lxplus. Edit the inputFileDir in doCondor.py to the directory where your samples are stored. Edit the currentDir in the doCondor.sh to match the directory that contains your LowPtTauMatcher.py and edit the line beneath 'export SCRAM_ARCH=<blah blah blah>', the one that begin 'cd' to be the path to the src directory of the CMSSW release in which you will call cmsenv.