complete beta version
rotation to all visible tau momentum points along Z in, once in that frame, ordering of pi in pT done, lead pT pi (in this vis tau momentum points along z frame) set to point along negative x axis, other pions and neutrinos rotated to live in this frame to (the toUse_local frame). Function to ensure that the phi associated pi 2, pi 3 is numerically between -pi/2 and pi/2 (these vectors will be due to physics, but this function ensures that we get out say -pi/4 as opposed to 7pi/4 in terms of the numerical value). Because this is not a rotation, we will not have to undo it.

The theta and phi of the visible tau vector before any rotation are saved, as is the phi associated with the lead pT phi in the visible tau momentum all points in Z frame (so the frame after one rotation is done) so that we can unrotate our vectors back to the lab frame at the end. Basically rotations take the form BA[orig vec], where A,B are rotation matrices. A is a function of the original visible tau momentum theta and phi, and B is a function of the lead pT phi in the frame we reach after rotating all the vectors by A (aka the visible tau momentum points in Z frame). 

To undo these rotations, you do: (A inv) (B inv) B A (vec), where A inv and B inv are the transpose matrices (transpose matrices are equivalent to the inverse matrix (see: https://math.stackexchange.com/questions/1936020/why-is-the-inverse-of-an-orthogonal-matrix-equal-to-its-transpose for a refresher as to why this is true, if needed). A inv and B inv are functions of the same things as A and B, respectively. Explicitly, A inv is a  function of original visible tau momentum theta and phi   and B inv is a function of the lead pT phi in the frame we reach after rotating all the vectors by A (aka the visible tau momentum points in Z frame).