Sort Taus And Rotate Them for Double Tau pT Gun sample
Note: all the stuff in here about "found an upsilon" or something else relating to upsilon is just code for: got both a tau and antitau (there is not in fact an upsilon here because this is a pT gun sample). Wording left over from the START_Upsilon code, which was the original. Because this is a tau/antitau pT gun and we want only cases where we recover both a tau and an antitau, it makes sense to leave in parts of code that guarantee we have both (which was what the upsilon stuff did in the original START_Upsilon). However, calling this guarantee that we have both a tau and an antitau an upsilon even in this double pT gun case is just a naming choice/choice made to avoid rewriting code unnecessarily. Do not let the reference to upsilons make you think there is a physical upsilon in this pT gun case, it is purely nomenclature for "have both a tau and an antitau" in this START_DoubleTauPtGun case.