START_Upsilon version that contains code to do the studies to determine the relative weights to be assigned to the inputs to the Figure of Merit labels that will be used for the DNN
At gen level, we check the effects of applying a Guassian smearing with sigma = 0.1 centered at the true value of a) nu pT normalized by tau mass, b) nu eta (theta in the local case), and c) nu phi (the inputs to our FOM in the DNN down the road). In each case, we then reconstruct the gen tau vector with everything that goes into it being exact except for the variable under study.  This is our attempt to numerically take a derivative, basically. We then examine the effects the smearing of each variable has on the resulting distribution in order to determine how to relatively weight the FOM inputs. We fit each smeared distribution with a Gaussian centered at the bin with highest occupancy in a window of +/- 2 times the RMS of the full smeared distribution. The Gaussian fits are not great, but as suggested by Greg, we do not worry about this and instead use the sigma from the (not great usually) Gaussian fit to define a window of +/- 4(5) times the Gaussian mean. We replot each distribution in this window around the mean and the take the RMS of the resulting distribution as the result of our derivative. We then compare the relative sizes of the RMS (our derivative proxy) for each label and weight them with weights proportional the RMS ratio^2. So for example, the RMS from the smeared local pT distribution is roughly twice that of the RMS from the smeared local theta distribution --> therefore the local pT receives a weight that is roughly 2^2 = 4 times as great as the weight given to the local theta in our calculation of the FOM in the training code down the road.